+++ scripts/postbuild.sh
@@ -1,0 +1,4 @@
+rm \
+  dist/pagination-astro/stub \
+  dist/pagination-vue/stub \
+  -rf